+++ Scraps.GUI/Bootstrapper.cs
@@ -46,7 +46,7 @@ namespace Scraps.GUI
                 Application.SetHighDpiMode(HighDpiMode.SystemAware);
                 Application.EnableVisualStyles();
                 Application.SetCompatibleTextRenderingDefault(false);
-                Application.Run(new MainForm());
+                Application.Run(new MainForm(args));
             }
         }